updated so it uses the same parsing
